==> working_nbs/nb2_text_extraction.ipynb <==
import os
import sys
import pandas as pd

sys.path.append('/home/workdir/document_classification')
from utils.data_extraction.extract_pdf import extract_text_from_pdfs


pdf_dir = "/home/ /workdir/document_classification/data/test_pdfs"
df_results = extract_text_from_pdfs(pdf_dir)

df_results.head(5)

df_results['status'].unique()

sum(df_results["status"] == "Success")

df_results[df_results["status"] == "Success"].head(10)


df_results[df_results["status"] == "Success"].iloc[25]['text']

df_results.to_csv(
    "/home/ /workdir/document_classification/data/v0/test/pdf_extraction_pypdf2_results.csv",
    index=False,
)

pdf_dir = "/home/ /workdir/document_classification/data/train_pdfs"
df_results_train = extract_text_from_pdfs(pdf_dir)


sum(df_results_train["status"] == "Success")

df_results_train["status"].unique()

df_results_train[df_results_train["status"] == "Failed: EOF marker not found"].head(10)


df_results_train.to_csv(
    "/home/ /workdir/document_classification/data/v0/train/pdf_extraction_pypdf2_results_train.csv",
    index=False,
)


from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(max_features=5000)
X_train = vectorizer.fit_transform(
    df_results_train[df_results_train["status"] == "Success"]["text"]
).toarray()

X_train[0]
