==> working_nbs/nb3_EDA.ipynb <==
import pandas as pd
import requests
import os
import ast
import sys
import matplotlib.pyplot as plt
import seaborn as sns
sys.path.append("/home/ /workdir/document_classification/utils")
from generic import get_random_sample,ast_df


excel_file = "/home/ /workdir/document_classification/data/DataSet.xlsx"
excel_sheets = pd.read_excel(excel_file, sheet_name=None)
for sheet_name, df in excel_sheets.items():
    print(sheet_name)
    globals()[sheet_name] = df

def prepare_filename_index(df):
    df = df.copy()
    df["doc_index"] = df.index
    df["filename"] = "doc_" + df["doc_index"].astype(str) + ".pdf"
    df.drop(columns=["doc_index"], inplace=True)
    return df


train_data = prepare_filename_index(train_data)
test_data = prepare_filename_index(test_data)

train_extracted_df = pd.read_csv('/home/ /workdir/document_classification/data/train/pdf_extraction_pypdf2_results_train.csv')
train_extracted_df = train_extracted_df[train_extracted_df["status"] == "Success"]
test_extracted_df = pd.read_csv('/home/ /workdir/document_classification/data/test/pdf_extraction_pypdf2_results_test.csv')
test_extracted_df = test_extracted_df[test_extracted_df["status"] == "Success"]


ast_df(train_extracted_df)
ast_df(test_extracted_df)

train_extracted_df = train_extracted_df.dropna()
test_extracted_df = test_extracted_df.dropna()

train_extracted_df.info()
test_extracted_df.info()

train_data

test_data

train_df = pd.merge(train_data, train_extracted_df, on="filename", how="inner")
test_df = pd.merge(test_data, test_extracted_df, on="filename", how="inner")
train_df = train_df[["filename", "text", "num_pages", "target_col"]]
test_df = test_df[["filename", "text", "num_pages", "target_col"]]

train_df

print("Dataset Info:")
print("-" * 50)
print(train_df.info())
print("\nMissing Values:")
print("-" * 50)
print(train_df.isnull().sum())



print("\nTrain: Target Class Distribution:")
print("-" * 50)
target_dist = train_df['target_col'].value_counts()
print(target_dist)
print("\nTrain: Percentage Distribution:")
print(target_dist/len(train_df)*100)

print("\nTest: Target Class Distribution:")
print("-" * 50)
target_dist = test_df['target_col'].value_counts()
print(target_dist)
print("\nTest: Percentage Distribution:")
print(target_dist/len(test_df)*100)


text, target = get_random_sample(train_df,class_name="others")

train_df['text_length'] = train_df['text'].str.len()
print("\nText Length Statistics:")
print("-" * 50)
print(train_df['text_length'].describe())
print("-" * 50)
print(f"90th percentile: {train_df['text_length'].quantile(0.90):.0f}")
print(f"95th percentile: {train_df['text_length'].quantile(0.95):.0f}")
print(f"99th percentile: {train_df['text_length'].quantile(0.99):.0f}")

print("\nNumber of Pages Statistics:")
print("-" * 50)
print(train_df['num_pages'].describe())
print("-" * 50)
print(f"90th percentile: {train_df['num_pages'].quantile(0.90):.0f}")
print(f"95th percentile: {train_df['num_pages'].quantile(0.95):.0f}")
print(f"99th percentile: {train_df['num_pages'].quantile(0.99):.0f}")

train_df.to_csv('/home/ /workdir/document_classification/data/train/uncleaned_train.csv', index=False)
test_df.to_csv('/home/ /workdir/document_classification/data/test/uncleaned_test.csv', index=False)


# Create a dictionary containing both dataframes
data_dict = {"train": train_df, "test": test_df}
# Save as pickle
pd.to_pickle(
    data_dict, "/home/ /workdir/document_classification/data/v0/combined_data.pkl"
)
